refactor: move settings card to widget
