+++ lib/features/settings/ui/view_model/settings_view_model.dart
@@ -1,5 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:hidroly/core/data/repositories/settings_repository_impl.dart';
 import 'package:hidroly/core/domain/enums/unit_systems.dart';
 import 'package:hidroly/core/providers/theme_provider.dart';
 import 'package:hidroly/core/providers/unit_system_provider.dart';
